reto-07 with protocols

+++ reto-07/eleloi/test/test_executers.py
@@ -7,7 +7,8 @@ from .sample_data import SAMPLE_DIR, none_dir, copy_dir, move_dir
 
 
 def _clean_sample_data_dir():
-    shutil.rmtree(SAMPLE_DIR)
+    if os.path.isdir(SAMPLE_DIR):
+        shutil.rmtree(SAMPLE_DIR)
     assert not os.path.isdir(SAMPLE_DIR)
 
 
@@ -27,8 +28,10 @@ def _fill_with_sample_files_with_extension(
 
 def test_none_executer_creates_directory():
     _clean_sample_data_dir()
-    executer = _executers.none
-    executer(none_dir.in_, none_dir.out, none_dir.filter_)
+    executer = _executers.NoneExecuter(
+        none_dir.in_, none_dir.out, none_dir.filter_
+    )
+    executer.run()
     assert os.path.isdir(SAMPLE_DIR)
 
 
@@ -39,8 +42,10 @@ def test_copy_executer_copy_all_files():
     _fill_with_sample_files_with_extension(copy_dir.in_, 2, "txt")
     assert len(os.listdir(copy_dir.in_)) == 5
     assert len(os.listdir(copy_dir.out)) == 0
-    executer = _executers.copy
-    executer(copy_dir.in_, copy_dir.out, copy_dir.filter_)
+    executer = _executers.CopyExecuter(
+        copy_dir.in_, copy_dir.out, copy_dir.filter_
+    )
+    executer.run()
     assert len(os.listdir(copy_dir.in_)) == 5
     assert len(os.listdir(copy_dir.out)) == 3
 
@@ -53,8 +58,10 @@ def test_copy_executer_copy_existing_files():
     _fill_with_sample_files_with_extension(copy_dir.in_, 2, "txt")
     assert len(os.listdir(copy_dir.in_)) == 5
     assert len(os.listdir(copy_dir.out)) == 1
-    executer = _executers.copy
-    executer(copy_dir.in_, copy_dir.out, copy_dir.filter_)
+    executer = _executers.CopyExecuter(
+        copy_dir.in_, copy_dir.out, copy_dir.filter_
+    )
+    executer.run()
     assert len(os.listdir(copy_dir.in_)) == 5
     assert len(os.listdir(copy_dir.out)) == 3
 
@@ -66,8 +73,10 @@ def test_move_executer_all_files():
     _fill_with_sample_files_with_extension(move_dir.in_, 2, "txt")
     assert len(os.listdir(move_dir.in_)) == 5
     assert len(os.listdir(move_dir.out)) == 0
-    executer = _executers.move
-    executer(move_dir.in_, move_dir.out, move_dir.filter_)
+    executer = _executers.MoveExecuter(
+        move_dir.in_, move_dir.out, move_dir.filter_
+    )
+    executer.run()
     assert len(os.listdir(move_dir.in_)) == 2
     assert len(os.listdir(move_dir.out)) == 3
 
@@ -80,7 +89,9 @@ def test_copy_executer_move_existing_files():
     _fill_with_sample_files_with_extension(move_dir.out, 1, "jpg")
     assert len(os.listdir(move_dir.in_)) == 5
     assert len(os.listdir(move_dir.out)) == 1
-    executer = _executers.move
-    executer(move_dir.in_, move_dir.out, move_dir.filter_)
+    executer = _executers.MoveExecuter(
+        move_dir.in_, move_dir.out, move_dir.filter_
+    )
+    executer.run()
     assert len(os.listdir(move_dir.in_)) == 2
     assert len(os.listdir(move_dir.out)) == 3
